Implement validation

+++ routes/api.php
@@ -14,32 +14,32 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
+Route::post('/user', function (Request $request) {
     return $request->user();
 });
 
 Route::post('/login', 'api\LoginController@login');
-Route::post('/logout', 'api\LoginController@logout');
+Route::middleware('auth:api')->post('/logout', 'api\LoginController@logout');
 
 //CRUD for cars
-Route::post('/cars','api\CarsController@getCars');
-Route::post('/car','api\CarsController@createCar');
-Route::put('/car/{id}','api\CarsController@updateCar');
-Route::delete('/car/{id}','api\CarsController@deleteCar');
+Route::middleware('auth:api')->post('/cars','api\CarsController@getCars');
+Route::middleware('auth:api')->post('/car','api\CarsController@createCar');
+Route::middleware('auth:api')->put('/car/{id}','api\CarsController@updateCar');
+Route::middleware('auth:api')->delete('/car/{id}','api\CarsController@deleteCar');
 
-Route::post('/dropdown-values','api\CarsController@getDropdownValues');
+Route::middleware('auth:api')->post('/dropdown-values','api\CarsController@getDropdownValues');
 
 //CRUD for manufacturers
-Route::post('/manufacturer','api\AdminController@createManufacturer');
-Route::put('/manufacturer/{id}','api\AdminController@updateManufacturer');
-Route::delete('/manufacturer/{id}','api\AdminController@deleteManufacturer');
+Route::middleware('auth:api')->post('/manufacturer','api\AdminController@createManufacturer');
+Route::middleware('auth:api')->put('/manufacturer/{id}','api\AdminController@updateManufacturer');
+Route::middleware('auth:api')->delete('/manufacturer/{id}','api\AdminController@deleteManufacturer');
 
 //CRUD for types
-Route::post('/type','api\AdminController@createType');
-Route::put('/type/{id}','api\AdminController@updateType');
-Route::delete('/type/{id}','api\AdminController@deleteType');
+Route::middleware('auth:api')->post('/type','api\AdminController@createType');
+Route::middleware('auth:api')->put('/type/{id}','api\AdminController@updateType');
+Route::middleware('auth:api')->delete('/type/{id}','api\AdminController@deleteType');
 
 //CRUD for colors
-Route::post('/color','api\AdminController@createColor');
-Route::put('/color/{id}','api\AdminController@updateColor');
-Route::delete('/color/{id}','api\AdminController@deleteColor');
+Route::middleware('auth:api')->post('/color','api\AdminController@createColor');
+Route::middleware('auth:api')->put('/color/{id}','api\AdminController@updateColor');
+Route::middleware('auth:api')->delete('/color/{id}','api\AdminController@deleteColor');
